Move isExecutionPatchIninitialResult to internal
